Differentiate between empty input and missing input
This changes Env.paramsEnv so that it can returns the
empty string when an empty string is submitted rather
than treating it as if no value was submitted.

The change is necessary in order to make it possible to
clear input fields that have a default value.

Obviously, this breaks backwards compatibility.